Написал тесты к NumberValidator и ObjectComparsion, рефакаторинг
Вынес классы Person, TsarRegistry и NumberValidator в отдельные файлы.
Перенес метод AreEqual в класс Person -> Equal
