+++ cs/HomeExercises/ObjectComparison.cs
@@ -21,16 +21,6 @@ namespace HomeExercises
 				options => options.Excluding(o => o.Path.EndsWith("Id"))
 				);
 
-			// Перепишите код на использование Fluent Assertions.
-			// Assert.AreEqual(actualTsar.Name, expectedTsar.Name);
-			// Assert.AreEqual(actualTsar.Age, expectedTsar.Age);
-			// Assert.AreEqual(actualTsar.Height, expectedTsar.Height);
-			// Assert.AreEqual(actualTsar.Weight, expectedTsar.Weight);
-			//
-			// Assert.AreEqual(expectedTsar.Parent!.Name, actualTsar.Parent!.Name);
-			// Assert.AreEqual(expectedTsar.Parent.Age, actualTsar.Parent.Age);
-			// Assert.AreEqual(expectedTsar.Parent.Height, actualTsar.Parent.Height);
-			// Assert.AreEqual(expectedTsar.Parent.Parent, actualTsar.Parent.Parent);
 		}
 
 		
@@ -48,48 +38,7 @@ namespace HomeExercises
 				new Person("Vasili III of Russia", 28, 170, 60, null));
 
 			// Какие недостатки у такого подхода? 
-			Assert.True(AreEqual(actualTsar, expectedTsar));
-		}
-
-		private bool AreEqual(Person? actual, Person? expected)
-		{
-			if (actual == expected) return true;
-			if (actual == null || expected == null) return false;
-			return
-				actual.Name == expected.Name
-				&& actual.Age == expected.Age
-				&& actual.Height == expected.Height
-				&& actual.Weight == expected.Weight
-				&& AreEqual(actual.Parent, expected.Parent);
-		}
-	}
-
-	public class TsarRegistry
-	{
-		public static Person GetCurrentTsar()
-		{
-			return new Person(
-				"Ivan IV The Terrible", 54, 170, 70,
-				new Person("Vasili III of Russia", 28, 170, 60, null));
-		}
-	}
-
-	public class Person
-	{
-		public static int IdCounter = 0;
-		public int Age, Height, Weight;
-		public string Name;
-		public Person? Parent;
-		public int Id;
-
-		public Person(string name, int age, int height, int weight, Person? parent)
-		{
-			Id = IdCounter++;
-			Name = name;
-			Age = age;
-			Height = height;
-			Weight = weight;
-			Parent = parent;
+			Assert.True(actualTsar.Equal(expectedTsar));
 		}
 	}
 }
